Setting up pipeline
- Changing base package of the project to io.github.danielmkraus
- Creating maven.yml pipeline
- Removing old travis ci configuration
- Configuring sonarcloud scan
