--- .travis.yml
@@ -1,16 +1,0 @@
-dist: trusty
-language: java
-jdk:
-  - openjdk11
-
-addons:
-  sonarcloud:
-    organization: "danielmkraus"
-
-script:
-  - mvn clean org.jacoco:jacoco-maven-plugin:prepare-agent verify org.jacoco:jacoco-maven-plugin:report -P coverage-per-test
-  - mvn sonar:sonar -P coverage-per-test
-
-cache:
-  directories:
-    - $HOME/.m2
