+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.danielmkraus</groupId>
+    <groupId>io.github.danielmkraus</groupId>
     <artifactId>jack-en-poy</artifactId>
     <version>1.0-SNAPSHOT</version>
     <properties>
@@ -14,8 +14,21 @@
 
         <version.junit>5.7.0</version.junit>
         <version.mockito>3.6.0</version.mockito>
-        <version.lombok>1.18.16</version.lombok>
-        <version.resteasy>4.5.8.Final</version.resteasy>
+        <version.lombok>1.18.30</version.lombok>
+        <version.resteasy>6.2.4.Final</version.resteasy>
+
+        <sonar.projectKey>jack-en-poy</sonar.projectKey>
+        <sonar.projectName>Jack en poy</sonar.projectName>
+        <sonar.links.homepage>https://github.com/danielmkraus/jack-en-poy
+        </sonar.links.homepage>
+        <sonar.links.ci>https://github.com/danielmkraus/jack-en-poy</sonar.links.ci>
+        <sonar.links.scm>https://github.com/danielmkraus/jack-en-poy</sonar.links.scm>
+        <sonar.links.issue>https://github.com/danielmkraus/jack-en-poy/issues
+        </sonar.links.issue>
+        <sonar.host.url>https://sonarcloud.io</sonar.host.url>
+        <sonar.organization>danielmkraus</sonar.organization>
+        <sonar.exclusions>target/**/*</sonar.exclusions>
+        <sonar.java.source>11</sonar.java.source>
     </properties>
 
     <dependencies>
@@ -32,22 +45,22 @@
         <dependency>
             <groupId>commons-cli</groupId>
             <artifactId>commons-cli</artifactId>
-            <version>1.4</version>
+            <version>1.5.0</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.datatype</groupId>
             <artifactId>jackson-datatype-jsr310</artifactId>
-            <version>2.11.3</version>
+            <version>2.15.3</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
-            <version>1.7.30</version>
+            <version>1.7.36</version>
         </dependency>
         <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
-            <version>1.2.3</version>
+            <version>1.4.14</version>
         </dependency>
 
         <dependency>
@@ -89,19 +102,19 @@
         <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
-            <version>2.8.6</version>
+            <version>2.10.1</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>com.fatboyindustrial.gson-javatime-serialisers</groupId>
             <artifactId>gson-javatime-serialisers</artifactId>
-            <version>1.1.1</version>
+            <version>1.1.2</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.awaitility</groupId>
             <artifactId>awaitility</artifactId>
-            <version>4.0.3</version>
+            <version>4.2.0</version>
             <scope>test</scope>
         </dependency>
 
@@ -117,7 +130,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
+                <version>3.11.0</version>
                 <configuration>
                     <annotationProcessorPaths>
                         <path>
@@ -131,12 +144,12 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-jar-plugin</artifactId>
-                <version>3.2.0</version>
+                <version>3.3.0</version>
                 <configuration>
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>org.danielmkraus.transfer.TransferServer</mainClass>
+                            <mainClass>io.github.danielmkraus.jackenpoy.Server</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
@@ -155,7 +168,7 @@
                     </descriptorRefs>
                     <archive>
                         <manifest>
-                            <mainClass>org.danielmkraus.jackenpoy.Server</mainClass>
+                            <mainClass>io.github.danielmkraus.jackenpoy.Server</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
