+++ src/main/java/io/github/danielmkraus/jackenpoy/domain/MatchResult.java
@@ -1,0 +1,7 @@
+package io.github.danielmkraus.jackenpoy.domain;
+
+public enum MatchResult {
+    WIN,
+    DRAW,
+    LOSE
+}
