+++ src/main/java/io/github/danielmkraus/jackenpoy/domain/User.java
@@ -1,0 +1,14 @@
+package io.github.danielmkraus.jackenpoy.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+    private String id;
+}
