+++ src/main/java/io/github/danielmkraus/jackenpoy/domain/player/Player.java
@@ -1,0 +1,7 @@
+package io.github.danielmkraus.jackenpoy.domain.player;
+
+import io.github.danielmkraus.jackenpoy.domain.Shape;
+
+public interface Player {
+    Shape play();
+}
