--- src/main/java/org/danielmkraus/jackenpoy/domain/Match.java
@@ -1,23 +1,0 @@
-package org.danielmkraus.jackenpoy.domain;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class Match {
-    private String id;
-    private final User user;
-    private final Shape shape;
-    private final Shape against;
-    private final LocalDateTime timestamp;
-
-    public MatchResult getResult(){
-        if (shape != null && against != null) {
-            return shape.against(against);
-        }
-        return null;
-    }
-}
