--- src/main/java/org/danielmkraus/jackenpoy/domain/MatchResult.java
@@ -1,7 +1,0 @@
-package org.danielmkraus.jackenpoy.domain;
-
-public enum MatchResult {
-    WIN,
-    DRAW,
-    LOSE
-}
