--- src/main/java/org/danielmkraus/jackenpoy/domain/User.java
@@ -1,14 +1,0 @@
-package org.danielmkraus.jackenpoy.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-    private String id;
-}
