--- src/main/java/org/danielmkraus/jackenpoy/domain/player/Player.java
@@ -1,7 +1,0 @@
-package org.danielmkraus.jackenpoy.domain.player;
-
-import org.danielmkraus.jackenpoy.domain.Shape;
-
-public interface Player {
-    Shape play();
-}
