+++ src/test/java/io/github/danielmkraus/jackenpoy/domain/player/FixedShapePlayerTest.java
@@ -1,0 +1,24 @@
+package io.github.danielmkraus.jackenpoy.domain.player;
+
+import io.github.danielmkraus.jackenpoy.domain.Shape;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNullPointerException;
+
+class FixedShapePlayerTest {
+    @ParameterizedTest
+    @EnumSource(Shape.class)
+    void fixed_shape_player_should_play_specified_shape(Shape shape){
+        var player = new FixedShapePlayer(shape);
+        Assertions.assertThat(player.play()).isEqualTo(shape);
+    }
+
+    @Test
+    void throw_null_pointer_exception_when_creates_a_fixed_shape_player_with_null(){
+        assertThatNullPointerException().isThrownBy(()-> new FixedShapePlayer(null));
+    }
+}
